Refactor existing optimize rules
