+++ arangod/Aql/Optimizer/Rule/CMakeLists.txt
@@ -1,5 +1,3 @@
-add_subdirectory(EnumeratePathsFilter)
-
 target_sources(arango_aql PRIVATE
   AsyncPrefetch.cpp
   BatchMaterializeDocuments.cpp
@@ -8,11 +6,13 @@ target_sources(arango_aql PRIVATE
   DistributeFilterCalcToCluster.cpp
   DistributeInCluster.cpp
   DistributeSortToCluster.cpp
+  EnumeratePathsFilter.cpp
   FuseFilters.cpp
   GeoIndex.cpp
   InlineSubqueries.cpp
   InterchangeAdjacentEnumerations.cpp
   JoinIndexNodes.cpp
+  LateDocumentMaterialization.cpp
   MaterializeIntoSeparateVariable.cpp
   MoveCalculationsDown.cpp
   MoveCalculationsUp.cpp
@@ -23,10 +23,11 @@ target_sources(arango_aql PRIVATE
   OptimizeProjections.cpp
   OptimizeSubqueries.cpp
   OptimizeTraversals.cpp
-  OptimizerRulesGraph.cpp
-  OptimizerRulesIResearchView.cpp
-  OptimizerRulesIndexNode.cpp
-  OptimizerRulesLateMaterializedCommon.cpp
+  HandleConstrainedSortInView.cpp
+  HandleViews.cpp
+  ImmutableSearchCondition.cpp
+  LateDocumentMaterializationArangoSearch.cpp
+  ScatterViewInCluster.cpp
   ParallelizeGather.cpp
   PropagateConstantAttributes.cpp
   PushDownLateMaterialization.cpp
@@ -45,6 +46,7 @@ target_sources(arango_aql PRIVATE
   RemoveUnnecessaryRemoteScatter.cpp
   ReplaceEntriesWithObjectIteration.cpp
   ReplaceEqualAttributeAccess.cpp
+  ReplaceLastAccessOnGraphPath.cpp
   ReplaceLikeWithRange.cpp
   ReplaceNearWithinFulltext.cpp
   ReplaceOrWithIn.cpp
