+++ arangod/Aql/Optimizer/Rule/EnumeratePathsFilter.cpp
@@ -26,8 +26,8 @@
 #include "Aql/ExecutionNode/EnumeratePathsNode.h"
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
+#include "Aql/Optimizer/Utils/EnumeratePathsFilterMatcher.h"
 #include "Aql/Optimizer.h"
-#include "Aql/Optimizer/Rule/EnumeratePathsFilter/EnumeratePathsFilterMatcher.h"
 #include "Containers/SmallVector.h"
 
 using namespace arangodb;
