--- arangod/Aql/Optimizer/Rule/EnumeratePathsFilter/CMakeLists.txt
@@ -1,3 +1,0 @@
-target_sources(arango_aql PRIVATE
-    EnumeratePathsFilterMatcher.cpp
-    EnumeratePathsFilter.cpp)
