+++ arangod/Aql/Optimizer/Rule/HandleViews.h
@@ -1,0 +1,43 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2024 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Andrey Abramov
+/// @author Vasiliy Nabatchikov
+////////////////////////////////////////////////////////////////////////////////
+
+#pragma once
+
+#include <memory>
+
+namespace arangodb::aql {
+
+class Optimizer;
+struct OptimizerRule;
+class ExecutionPlan;
+
+}  // namespace arangodb::aql
+
+namespace arangodb::iresearch {
+
+void handleViewsRule(aql::Optimizer* opt,
+                     std::unique_ptr<aql::ExecutionPlan> plan,
+                     aql::OptimizerRule const& rule);
+
+}  // namespace arangodb::iresearch
