+++ arangod/Aql/Optimizer/Rule/LateDocumentMaterialization.cpp
@@ -21,6 +21,8 @@
 /// @author Yuriy Popov
 ////////////////////////////////////////////////////////////////////////////////
 
+#include "LateDocumentMaterialization.h"
+
 #include "Aql/Ast.h"
 #include "Aql/CalculationNodeVarFinder.h"
 #include "Aql/Collection.h"
@@ -33,14 +35,13 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/Expression.h"
 #include "Aql/Optimizer.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIndexNode.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesLateMaterializedCommon.h"
+#include "Aql/Optimizer/Utils/LateMaterializedCommon.h"
 #include "Aql/QueryContext.h"
 #include "Basics/AttributeNameParser.h"
 #include "Cluster/ServerState.h"
 #include "Indexes/Index.h"
 
-using namespace arangodb::aql;
+namespace arangodb::aql {
 
 namespace {
 
@@ -93,7 +94,7 @@ bool processCalculationNode(
 
 }  // namespace
 
-void arangodb::aql::lateDocumentMaterializationRule(
+void lateDocumentMaterializationRule(
     Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
     OptimizerRule const& rule) {
   auto modified = false;
@@ -372,3 +373,5 @@ void arangodb::aql::lateDocumentMaterializationRule(
     }
   }
 }
+
+}  // namespace arangodb::aql
