+++ arangod/Aql/Optimizer/Utils/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(arango_aql PRIVATE
   AccessesCollectionVariable.cpp
+  EnumeratePathsFilterMatcher.cpp
+  LateMaterializedCommon.cpp
   OptimizeTraversalPathVariable.cpp
 )
