+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -30,12 +30,13 @@
 #include "Aql/Optimizer/Rule/DistributeFilterCalcToCluster.h"
 #include "Aql/Optimizer/Rule/DistributeInCluster.h"
 #include "Aql/Optimizer/Rule/DistributeSortToCluster.h"
-#include "Aql/Optimizer/Rule/EnumeratePathsFilter/EnumeratePathsFilter.h"
+#include "Aql/Optimizer/Rule/EnumeratePathsFilter.h"
 #include "Aql/Optimizer/Rule/FuseFilters.h"
 #include "Aql/Optimizer/Rule/GeoIndex.h"
 #include "Aql/Optimizer/Rule/InlineSubqueries.h"
 #include "Aql/Optimizer/Rule/InterchangeAdjacentEnumerations.h"
 #include "Aql/Optimizer/Rule/JoinIndexNodes.h"
+#include "Aql/Optimizer/Rule/LateDocumentMaterialization.h"
 #include "Aql/Optimizer/Rule/MaterializeIntoSeparateVariable.h"
 #include "Aql/Optimizer/Rule/MoveCalculationsDown.h"
 #include "Aql/Optimizer/Rule/MoveCalculationsUp.h"
@@ -46,9 +47,11 @@
 #include "Aql/Optimizer/Rule/OptimizeProjections.h"
 #include "Aql/Optimizer/Rule/OptimizeSubqueries.h"
 #include "Aql/Optimizer/Rule/OptimizeTraversals.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesGraph.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIResearchView.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIndexNode.h"
+#include "Aql/Optimizer/Rule/HandleConstrainedSortInView.h"
+#include "Aql/Optimizer/Rule/HandleViews.h"
+#include "Aql/Optimizer/Rule/ImmutableSearchCondition.h"
+#include "Aql/Optimizer/Rule/LateDocumentMaterializationArangoSearch.h"
+#include "Aql/Optimizer/Rule/ScatterViewInCluster.h"
 #include "Aql/Optimizer/Rule/ParallelizeGather.h"
 #include "Aql/Optimizer/Rule/PropagateConstantAttributes.h"
 #include "Aql/Optimizer/Rule/PushDownLateMaterialization.h"
@@ -67,6 +70,7 @@
 #include "Aql/Optimizer/Rule/RemoveUnnecessaryRemoteScatter.h"
 #include "Aql/Optimizer/Rule/ReplaceEntriesWithObjectIteration.h"
 #include "Aql/Optimizer/Rule/ReplaceEqualAttributeAccess.h"
+#include "Aql/Optimizer/Rule/ReplaceLastAccessOnGraphPath.h"
 #include "Aql/Optimizer/Rule/ReplaceLikeWithRange.h"
 #include "Aql/Optimizer/Rule/ReplaceNearWithinFulltext.h"
 #include "Aql/Optimizer/Rule/ReplaceOrWithIn.h"
